+++ src/main/java/online/kingdomkeys/kingdomkeys/entity/magic/SparkEntity.java
@@ -10,6 +10,7 @@ import net.minecraft.world.entity.projectile.ThrowableProjectile;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
+import online.kingdomkeys.kingdomkeys.damagesource.KKDamageTypes;
 import online.kingdomkeys.kingdomkeys.data.WorldData;
 import online.kingdomkeys.kingdomkeys.entity.ModEntities;
 import online.kingdomkeys.kingdomkeys.lib.Party;
@@ -36,6 +37,7 @@ public class SparkEntity extends BaseMagicProjectile {
 		super(ModEntities.TYPE_SPARK.get(), owner, level);
 		this.dmgMult = dmgMult;
 		this.blocksBuilding = true;
+		this.damageType = KKDamageTypes.LIGHTNING;
 		setIndex(index);
 	}